requirements.txt
asgiref==3.8.1
Django>=4.2
djangorestframework==3.15.2
django-jazzmin==3.0.0
gunicorn==23.0.0
packaging==24.1
sqlparse==0.5.1
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
django-resized==1.0.2
pillow==10.4.0
python-dateutil==2.9.0.post0
six==1.16.0
django-admin-soft-dashboard==1.0.14
str2bool==1.1